[Green/Refactor][Whiteboard/Element]: Se codifica lo mínimo necesario para que las pruebas pasen, y se mejoran aspectos de claridad y calidad del código fuente.

+++ Collaborative Board/Domain/Password.cs
@@ -9,15 +9,11 @@ namespace Domain
     {
         private static readonly ReadOnlyCollection<char> allowedCharacters =
             Array.AsReadOnly("ABCDEFGHIJKLMNÑOPQRSTUVWXYZabcdefghijklmnñopqrstuvwxyz0123456789".ToCharArray());
-
-        private const int minimumLength = 6;
-
-        private const int maximumLength = 20;
-
-        private const int lengthOfGeneratedPassword = 10;
+        private const byte minimumLength = 6;
+        private const byte maximumLength = 20;
+        private const byte lengthOfGeneratedPassword = 10;
 
         private string passwordValue;
-
         internal string PasswordValue
         {
             get { return passwordValue; }
